refactor: DTO 클래스들의 패키지를 변경
- 중첩 클래스를 사용하며 현재처럼 하나의 톱레벨 dto 클래스만 존재하는 경우에는 하위 패키지에 두는 것보다는 꺼내두는게 더 적절하다고 생각했다
- 만약 Request, Response DTO 를 나누게 된다면 하위 패키지를 만드는 것을 고려 해 볼 수 있을 듯 하다
