+++ src/main/java/com/prgrms/boardjpa/post/PostRestController.java
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.prgrms.boardjpa.commons.api.SuccessResponse;
-import com.prgrms.boardjpa.post.dto.PostDto;
 
 @RequestMapping("/api/posts")
 @RestController
