+++ src/main/java/com/prgrms/boardjpa/post/PostService.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.prgrms.boardjpa.commons.exception.NotExistException;
-import com.prgrms.boardjpa.post.dto.PostDto;
 import com.prgrms.boardjpa.user.domain.User;
 import com.prgrms.boardjpa.user.UserService;
 import com.prgrms.boardjpa.user.exception.AuthorizationFailException;
